requirements.txt
fastapi
uvicorn
openai
sentence_transformers
pymupdf
numpy
pydantic
psycopg2-binary
torch
torch_geometric
python-multipart
python-dotenv
